Made a few big changes: 1) Specified that autoconf should interpret this as a c++ program. 2) Added a compiler flag that specifies we're using the c++0x (a.k.a. c++11) standard. This may break compatibility with compilers other than g++. Sorry about that, but I don't know how to do this in a compiler-neutral way. 3) Converted most integers from uint32_t (and similar for other bit widths), which is standard but optional, to uint_least32_t or uint_fast32_t (and corresponding for other bit widths) since those are not optional. Some integers in the font-related classes are still Irrlicht u32's. 4) Reduced the widths of lots of integers from 32 to 16 bits. This greatly sped up the game when running on the Raspberry Pi in windowed mode but for some reason has no noticeable effect when running fullscreen. Watch out for bugs that may have been introduced; I fixed all I found but there may be more. 5) Rewrote the texture creation and drawing code to look good on R Pi. Not sure if that was in previous commit or not; anyway it's there now.

+++ src/AI.h
@@ -1,11 +1,11 @@
 /**
  * Copyright © 2013 James Dearing.
  * This file is part of Cybrinth.
- * 
+ *
  * Cybrinth is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
- * 
+ *
  * Cybrinth is distributed in the hope that it will be fun, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Affero General Public License along with Cybrinth. If not, see <http://www.gnu.org/licenses/>.
 */
 
@@ -25,23 +25,23 @@ class AI {
 		AI();
 		/** Default destructor */
 		virtual ~AI();
-		void setPlayer( uint8_t newPlayer );
-		void setup( MazeCell ** newMaze, uint8_t newCols, uint8_t newRows,  GameManager * newGM ); //Provides the AI with whatever info it needs to work.
+		void setPlayer( uint_least8_t newPlayer );
+		void setup( MazeCell ** newMaze, uint_least8_t newCols, uint_least8_t newRows,  GameManager * newGM ); //Provides the AI with whatever info it needs to work.
 		void move(); //Needs to call GameManager's movePlayerOnX and movePlayerOnY functions.
 		void allKeysFound();
-		uint32_t lastTimeMoved;
-		uint8_t getPlayer();
-		uint32_t movementDelay; //How long to delay between movements
+		uint_least16_t lastTimeMoved;
+		uint_least8_t getPlayer();
+		uint_least16_t movementDelay; //How long to delay between movements
 	protected:
 	private:
 		GameManager * gm;
-		uint8_t controlsPlayer;
-		std::vector< irr::core::position2d< uint8_t > > cellsVisited;
-		std::vector< irr::core::position2d< uint8_t > > pathTaken;
-		std::vector< std::vector< irr::core::position2d< uint8_t > > > pathsToLockedCells;
-		bool alreadyVisited( irr::core::position2d< uint8_t > position );
-		uint8_t cols;
-		uint8_t rows;
+		uint_least8_t controlsPlayer;
+		std::vector< irr::core::position2d< uint_least8_t > > cellsVisited;
+		std::vector< irr::core::position2d< uint_least8_t > > pathTaken;
+		std::vector< std::vector< irr::core::position2d< uint_least8_t > > > pathsToLockedCells;
+		bool alreadyVisited( irr::core::position2d< uint_least8_t > position );
+		uint_least8_t cols;
+		uint_least8_t rows;
 		MazeCell ** maze;
 };
