+++ src/Goal.h
@@ -20,9 +20,9 @@ class Goal : public Object {
 	public:
 		Goal();
 		virtual ~Goal();
-		void draw( irr::video::IVideoDriver* driver, uint32_t width, uint32_t height );
+		void draw( irr::video::IVideoDriver* driver, uint_least16_t width, uint_least16_t height );
 		void loadTexture( irr::video::IVideoDriver* driver );
-		void loadTexture( irr::video::IVideoDriver* driver, uint32_t size );
+		void loadTexture( irr::video::IVideoDriver* driver, uint_least16_t size );
 	protected:
 	private:
 };
