+++ src/MazeManager.cpp
@@ -8,7 +8,7 @@
 using namespace std;
 
 //Recursively searches the maze to see if it can get from start to end
-bool MazeManager::canGetTo( uint8_t startX, uint8_t startY, uint8_t goalX, uint8_t goalY ) {
+bool MazeManager::canGetTo( uint_least8_t startX, uint_least8_t startY, uint_least8_t goalX, uint_least8_t goalY ) {
 	bool found = false;
 	if( gameManager->getDebugStatus() ) {
 		wcout << L"Searching for a way from " << static_cast<unsigned int>( startX ) << L"x" << static_cast<unsigned int>( startY ) << L" to " << static_cast<unsigned int>( goalX ) << L"x" << static_cast<unsigned int>( goalY ) << endl;
@@ -39,14 +39,14 @@ bool MazeManager::canGetTo( uint8_t startX, uint8_t startY, uint8_t goalX, uint8
 }
 
 //Iterates through all collectables, calls canGetTo on their locations
-bool MazeManager::canGetToAllCollectables( uint8_t startX, uint8_t startY ) {
+bool MazeManager::canGetToAllCollectables( uint_least8_t startX, uint_least8_t startY ) {
 	bool result = true;
 
-	for( uint32_t i = 0; ( i < gameManager->stuff.size() && result == true ); i++ ) {
+	for( uint_fast16_t i = 0; ( i < gameManager->stuff.size() && result == true ); i++ ) {
 
 		//Do this here because canGetTo() uses the visited variable
-		for( uint8_t x = 0; x < cols; x++ ) {
-			for( uint8_t y = 0; y < rows; y++ ) {
+		for( uint_fast8_t x = 0; x < cols; x++ ) {
+			for( uint_fast8_t y = 0; y < rows; y++ ) {
 				maze[x][y].visited = false;
 			}
 		}
@@ -61,12 +61,12 @@ bool MazeManager::canGetToAllCollectables( uint8_t startX, uint8_t startY ) {
 	return result;
 }
 
-void MazeManager::draw( video::IVideoDriver* driver, uint32_t cellWidth, uint32_t cellHeight ) {
+void MazeManager::draw( video::IVideoDriver* driver, uint_least16_t cellWidth, uint_least16_t cellHeight ) {
 	video::SColor wallColor = WHITE;
 	video::SColor lockColor = BROWN;
 
-	for( uint8_t x = 0; x < cols; x++ ) {
-		for( uint8_t y = 0; y < rows; y++ ) {
+	for( uint_fast8_t x = 0; x < cols; x++ ) {
+		for( uint_fast8_t y = 0; y < rows; y++ ) {
 			if( maze[x][y].visible ) {
 				if( maze[x][y].getTop() == 'w' ) {
 					driver->draw2DLine( core::position2d< irr::s32 >( cellWidth * x, cellHeight * y ), core::position2d< irr::s32 >(cellWidth * ( x + 1 ), cellHeight * y ), wallColor );
@@ -93,14 +93,14 @@ void MazeManager::draw( video::IVideoDriver* driver, uint32_t cellWidth, uint32_
 	}
 }
 
-bool MazeManager::existsAnythingAt( uint8_t x, uint8_t y ) {
+bool MazeManager::existsAnythingAt( uint_least8_t x, uint_least8_t y ) {
 	bool result = false;
 
 	if( gameManager->goal.getX() == x && gameManager->goal.getY() == y ) {
 		result = true;
 	}
 
-	for( uint32_t i = 0; ( result == false && i < gameManager->stuff.size() ); i++ ) {
+	for( uint_fast16_t i = 0; ( result == false && i < gameManager->stuff.size() ); i++ ) {
 		if( gameManager->stuff[ i ].getX() == x && gameManager->stuff[ i ].getY() == y ) {
 			result = true;
 		}
@@ -153,12 +153,12 @@ void MazeManager::makeRandomLevel() {
 	gameManager->drawAll();
 
 	srand( gameManager->randomSeed ); //randomSeed is set either by resetThings() or by loadFromFile()
-	uint8_t tempCols = rand() % 28 + 2;
-	uint8_t tempRows = tempCols + ( rand() % 5 );
+	uint_least8_t tempCols = rand() % 28 + 2;
+	uint_least8_t tempRows = tempCols + ( rand() % 5 );
 	resizeMaze( tempCols, tempRows );
 
-	for( uint8_t x = 0; x < cols; x++ ) {
-		for( uint8_t y = 0; y < rows; y++ ) {
+	for( uint_fast8_t x = 0; x < cols; x++ ) {
+		for( uint_fast8_t y = 0; y < rows; y++ ) {
 			maze[x][y].setTop( 'w' );
 			maze[x][y].setLeft( 'w' );
 			maze[x][y].setRight( 'n' );
@@ -167,7 +167,7 @@ void MazeManager::makeRandomLevel() {
 		}
 	}
 
-	for( uint8_t p = 0; p < gameManager->numPlayers; p++ ) {
+	for( uint_fast8_t p = 0; p < gameManager->numPlayers; p++ ) {
 		gameManager->playerStart[p].reset();
 	}
 
@@ -184,34 +184,34 @@ void MazeManager::makeRandomLevel() {
 	recurseRandom( gameManager->goal.getX(), gameManager->goal.getY(), 0, 0 ); //Start recursion from the gameManager->goal; for some reason that makes the mazes harder than if we started recursion from the player's starting point.
 
 	//Add walls at maze borders
-	for( uint8_t x = 0; x < cols; x++ ) {
+	for( uint_fast8_t x = 0; x < cols; x++ ) {
 		maze[x][0].setTop( 'w' );
 		maze[x][rows-1].setBottom( 'w' );
 	}
 
-	for( uint8_t y = 0; y < rows; y++ ) {
+	for( uint_fast8_t y = 0; y < rows; y++ ) {
 		maze[0][y].setLeft( 'w' );
 		maze[cols-1][y].setRight( 'w' );
 	}
 
-	for( uint8_t x = 1; x < cols; x++ ) {
-		for( uint8_t y = 0; y < rows; y++ ) {
+	for( uint_fast8_t x = 1; x < cols; x++ ) {
+		for( uint_fast8_t y = 0; y < rows; y++ ) {
 			maze[x-1][y].setRight( maze[x][y].getLeft() );
 		}
 	}
 
-	for( uint8_t x = 0; x < cols; x++ ) {
-		for( uint8_t y = 1; y < rows; y++ ) {
+	for( uint_fast8_t x = 0; x < cols; x++ ) {
+		for( uint_fast8_t y = 1; y < rows; y++ ) {
 			maze[x][y-1].setBottom( maze[x][y].getTop() );
 		}
 	}
 
 	//Find all dead ends. I'm sure it would be more efficient to do this during maze generation rather than going back through afterward, but I can't be bothered with that now.
-	vector<uint8_t> deadEndsX;
-	vector<uint8_t> deadEndsY;
+	vector<uint_least8_t> deadEndsX;
+	vector<uint_least8_t> deadEndsY;
 
-	for( uint8_t x = 0; x < cols; x++ ) {
-		for( uint8_t y = 0; y < rows; y++ ) {
+	for( uint_fast8_t x = 0; x < cols; x++ ) {
+		for( uint_fast8_t y = 0; y < rows; y++ ) {
 			if( maze[x][y].isDeadEnd() ) {
 				deadEndsX.push_back( x );
 				deadEndsY.push_back( y );
@@ -220,8 +220,8 @@ void MazeManager::makeRandomLevel() {
 	}
 
 	//Remove player starts from list of dead ends
-	for( uint8_t p = 0; p < gameManager->numPlayers; p++ ) {
-		for( uint32_t i = 0; i < deadEndsX.size(); i++ ) {
+	for( uint_fast8_t p = 0; p < gameManager->numPlayers; p++ ) {
+		for( uint_fast16_t i = 0; i < deadEndsX.size(); i++ ) {
 			if( gameManager->playerStart[p].getX() == deadEndsX[i] && gameManager->playerStart[p].getY() == deadEndsY[i] ) {
 				deadEndsX.erase( deadEndsX.begin() + i );
 				deadEndsY.erase( deadEndsY.begin() + i );
@@ -230,7 +230,7 @@ void MazeManager::makeRandomLevel() {
 	}
 
 	//Remove gameManager->goal from list of dead ends
-	for( uint32_t i = 0; i < deadEndsX.size(); i++ ) {
+	for( uint_fast16_t i = 0; i < deadEndsX.size(); i++ ) {
 		if( gameManager->goal.getX() == deadEndsX[i] && gameManager->goal.getY() == deadEndsY[i] ) {
 			deadEndsX.erase( deadEndsX.begin() + i );
 			deadEndsY.erase( deadEndsY.begin() + i );
@@ -241,21 +241,21 @@ void MazeManager::makeRandomLevel() {
 		gameManager->numLocks = deadEndsX.size();
 	}
 
-	uint8_t numKeys = gameManager->numLocks;
+	uint_least8_t numKeys = gameManager->numLocks;
 
 	//Place keys in dead ends
-	vector<uint8_t> keyPlaceX;
-	vector<uint8_t> keyPlaceY;
+	vector<uint_least8_t> keyPlaceX;
+	vector<uint_least8_t> keyPlaceY;
 
-	for( uint8_t k = 0; k < numKeys; k++ ) {
-		vector<uint8_t> chosenPlaces;
+	for( uint_fast8_t k = 0; k < numKeys; k++ ) {
+		vector<uint_least8_t> chosenPlaces;
 
 		if( deadEndsX.size() == 0 ) {
 			deadEndsX.push_back( gameManager->playerStart[0].getX() );
 			deadEndsY.push_back( gameManager->playerStart[0].getY() );
 		}
 		//Pick one of the dead ends randomly.
-		uint8_t chosen = rand() % deadEndsX.size();
+		uint_least8_t chosen = rand() % deadEndsX.size();
 
 		//Finally, create a key and put it there.
 		Collectable temp;
@@ -270,7 +270,7 @@ void MazeManager::makeRandomLevel() {
 		deadEndsY.erase( deadEndsY.begin() + chosen );
 	}
 
-	for( uint8_t p = 0; p < gameManager->numPlayers; p++ ) {
+	for( uint_fast8_t p = 0; p < gameManager->numPlayers; p++ ) {
 		gameManager->player[p].setPos( gameManager->playerStart[p].getX(), gameManager->playerStart[p].getY() );
 	}
 
@@ -287,11 +287,11 @@ void MazeManager::makeRandomLevel() {
 			maze[gameManager->goal.getX() + 1][gameManager->goal.getY()].setLeft( 'l' );
 		}
 
-		uint8_t numLocksPlaced = 1;
+		uint_fast8_t numLocksPlaced = 1;
 
 		while( gameManager->device->run() != false && numLocksPlaced < gameManager->numLocks && gameManager->timer->getTime() < gameManager->timeStartedLoading + gameManager->loadingDelay ) {
-			uint8_t tempX = rand() % cols;
-			uint8_t tempY = rand() % rows;
+			uint_least8_t tempX = rand() % cols;
+			uint_least8_t tempY = rand() % rows;
 
 			if( maze[tempX][tempY].getTop() == 'n' ) {
 				maze[tempX][tempY].setTop( 'l' );
@@ -327,7 +327,7 @@ void MazeManager::makeRandomLevel() {
 		if( numLocksPlaced < gameManager->numLocks ) {
 			int keysToRemove = gameManager->numLocks - numLocksPlaced;
 
-			for( uint32_t i = 0; ( i < gameManager->stuff.size() && keysToRemove > 0 ); i++ ) {
+			for( uint_fast16_t i = 0; ( i < gameManager->stuff.size() && keysToRemove > 0 ); i++ ) {
 				if( gameManager->getDebugStatus() ) {
 					wcout << L"keysToRemove: " << static_cast<unsigned int>( keysToRemove ) << endl;
 				}
@@ -343,20 +343,20 @@ void MazeManager::makeRandomLevel() {
 		numKeys = gameManager->numLocks = numLocksPlaced;
 	}
 
-	for( uint8_t x = 0; x < cols; x++ ) {
-		for( uint8_t y = 0; y < rows; y++ ) {
+	for( uint_fast8_t x = 0; x < cols; x++ ) {
+		for( uint_fast8_t y = 0; y < rows; y++ ) {
 			maze[x][y].visited = false;
 		}
 	}
 
-	for( uint8_t p = 0; p < gameManager->numPlayers; p++ ) {
+	for( uint_fast8_t p = 0; p < gameManager->numPlayers; p++ ) {
 		maze[ gameManager->playerStart[ p ].getX()][ gameManager->playerStart[ p ].getY()].visited = true;
 		maze[ gameManager->playerStart[ p ].getX()][ gameManager->playerStart[ p ].getY()].setVisitorColor( gameManager->player[ p ].getColorTwo() );
 	}
 
 	//Set up bots;
 	if( gameManager->numBots > 0 ) {
-		for( uint8_t i = 0; i < gameManager->numBots; i++ ) {
+		for( uint_fast8_t i = 0; i < gameManager->numBots; i++ ) {
 			gameManager->bot[ i ].setup( maze, cols, rows, gameManager );
 		}
 	}
@@ -371,7 +371,7 @@ MazeManager::MazeManager()
 
 MazeManager::~MazeManager()
 {
-	for( uint8_t i = 0 ; i < cols ; i++ ) {
+	for( uint_fast8_t i = 0 ; i < cols ; i++ ) {
 		delete [] maze[ i ];
 	}
 
@@ -379,14 +379,14 @@ MazeManager::~MazeManager()
 }
 
 //Generates the maze recursively
-void MazeManager::recurseRandom( uint8_t x, uint8_t y, uint16_t depth, uint16_t numSoFar ) {
+void MazeManager::recurseRandom( uint_least8_t x, uint_least8_t y, uint_least16_t depth, uint_least16_t numSoFar ) {
 	gameManager->device->run();
 	gameManager->drawAll();
 
 	maze[x][y].visited = true;
 	maze[x][y].id = numSoFar;
 
-	for( uint8_t p = 0; p < gameManager->numPlayers; p++ ) {
+	for( uint_fast8_t p = 0; p < gameManager->numPlayers; p++ ) {
 		if( depth >= gameManager->playerStart[p].distanceFromExit ) {
 			gameManager->playerStart[p].setPos( x, y );
 			gameManager->playerStart[p].distanceFromExit = depth;
@@ -453,9 +453,9 @@ void MazeManager::recurseRandom( uint8_t x, uint8_t y, uint16_t depth, uint16_t 
 }
 
 //Creates a new maze of the desired size, copies from old maze to new, then deletes old maze
-void MazeManager::resizeMaze( uint8_t newCols, uint8_t newRows ) {
-	uint8_t oldCols = cols;
-	uint8_t oldRows = rows;
+void MazeManager::resizeMaze( uint_least8_t newCols, uint_least8_t newRows ) {
+	uint_least8_t oldCols = cols;
+	uint_least8_t oldRows = rows;
 
 	if( gameManager->getDebugStatus() && ( newCols < oldCols || newRows < oldRows ) ) {
 		wcerr << L"Warning: New maze size smaller than old in some dimension. newCols: " << static_cast<unsigned int>( newCols ) << L" oldCols: " << static_cast<unsigned int>( oldCols ) << L" newRows: " << static_cast<unsigned int>( newRows ) << L" oldRows: " << static_cast<unsigned int>( oldRows ) << endl;
@@ -463,11 +463,11 @@ void MazeManager::resizeMaze( uint8_t newCols, uint8_t newRows ) {
 
 	MazeCell** temp = new MazeCell *[newCols];
 
-	for( int i = 0 ; i < newCols ; i++ ) {
+	for( uint_fast8_t i = 0 ; i < newCols ; i++ ) {
 		temp[i] = new MazeCell[newRows];
 	}
 
-	uint8_t colsToCopy;
+	uint_least8_t colsToCopy;
 
 	if( newCols > oldCols ) {
 		colsToCopy = oldCols;
@@ -475,7 +475,7 @@ void MazeManager::resizeMaze( uint8_t newCols, uint8_t newRows ) {
 		colsToCopy = newCols;
 	}
 
-	uint8_t rowsToCopy;
+	uint_least8_t rowsToCopy;
 
 	if( newRows > oldRows ) {
 		rowsToCopy = oldRows;
@@ -483,13 +483,13 @@ void MazeManager::resizeMaze( uint8_t newCols, uint8_t newRows ) {
 		rowsToCopy = newRows;
 	}
 
-	for( uint8_t x = 0; x < colsToCopy; x++ ) {
-		for( uint8_t y = 0; y < rowsToCopy; y++ ) {
+	for( uint_fast8_t x = 0; x < colsToCopy; x++ ) {
+		for( uint_fast8_t y = 0; y < rowsToCopy; y++ ) {
 			temp[x][y] = maze[x][y];
 		}
 	}
 
-	for( uint8_t i = 0 ; i < oldCols ; i++ ) {
+	for( uint_fast8_t i = 0 ; i < oldCols ; i++ ) {
 		delete [] maze[ i ];
 	}
