+++ src/MazeManager.h
@@ -18,34 +18,34 @@ class MazeManager
 		MazeManager();
 		/** Default destructor */
 		virtual ~MazeManager();
-		
+
 		MazeCell** maze;
-		
-		bool canGetTo( uint8_t startX, uint8_t startY, uint8_t goalX, uint8_t goalY );
-		bool canGetToAllCollectables( uint8_t startX, uint8_t startY );
-		
-		void draw( irr::video::IVideoDriver* driver, uint32_t cellWidth, uint32_t cellHeight );
-		
-		bool existsAnythingAt( uint8_t x, uint8_t y );
-		
+
+		bool canGetTo( uint_least8_t startX, uint_least8_t startY, uint_least8_t goalX, uint_least8_t goalY );
+		bool canGetToAllCollectables( uint_least8_t startX, uint_least8_t startY );
+
+		void draw( irr::video::IVideoDriver* driver, uint_least16_t cellWidth, uint_least16_t cellHeight );
+
+		bool existsAnythingAt( uint_least8_t x, uint_least8_t y );
+
 		bool loadFromFile();
 		bool loadFromFile( boost::filesystem::path src );
-		
+
 		void makeRandomLevel();
-		
-		void recurseRandom( uint8_t x, uint8_t y, uint16_t depth, uint16_t numSoFar );
-		void resizeMaze( uint8_t newCols, uint8_t newRows );
-		
+
+		void recurseRandom( uint_least8_t x, uint_least8_t y, uint_least16_t depth, uint_least16_t numSoFar );
+		void resizeMaze( uint_least8_t newCols, uint_least8_t newRows );
+
 		bool saveToFile();
 		bool saveToFile( boost::filesystem::path dest );
-		
+
 		void setGameManager( GameManager* newGM );
-		
-		uint8_t cols;
-		
+
+		uint_least8_t cols;
+
 		GameManager* gameManager;
-		
-		uint8_t rows;
+
+		uint_least8_t rows;
 	protected:
 	private:
 };
